fixed filter of scores

+++ src/testing_system/models/test_results_request.py
@@ -10,9 +10,9 @@ from testing_system.models.test_results import OrderingEnum
 class TestResultsRequest(BaseModel):
     only_max_result: bool = None
     search_prefix: str = None
-    upper_bound: int = None
-    lower_bound: int = None
-    score_equals: int = None
+    upper_bound: float = None
+    lower_bound: float = None
+    score_equals: float = None
     date_from: Date = None
     date_to: Date = None
     ordering: Optional[OrderingEnum] = None
